Issue #3942
PR #3971

+++ monkey/monkey_island/cc/resources/propagation_credentials.py
@@ -35,7 +35,22 @@ class PropagationCredentials(AbstractResource):
     @auth_token_required
     @roles_accepted(AccountRole.ISLAND_INTERFACE.name)
     def put(self, collection=None):
-        credentials = [Credentials(**c) for c in request.json]
+        credentials = []
+        errors = []
+        for index, credential_pair in enumerate(request.json):
+            try:
+                credentials.append(Credentials(**credential_pair))
+            except (TypeError, ValueError) as err:
+                errors.append(
+                    {
+                        "message": f"{str(err)}",
+                        "index": index,
+                    }
+                )
+
+        if errors:
+            return {"errors": errors}, HTTPStatus.BAD_REQUEST
+
         if collection == _configured_collection:
             self._credentials_repository.remove_configured_credentials()
             self._credentials_repository.save_configured_credentials(credentials)
